@@ -16,10 +16,8 @@ class Alerts extends AlertsConfig
 {
     /**
      * Template to use for HTML output.
-     *
-     * @var string
      */
-    public $template = 'Tatter\Alerts\Views\Bootstrap4';
+    public string $template = 'Tatter\Alerts\Views\Bootstrap4';
 
     /**
      * Mapping of Session keys to their CSS classes.
@@ -28,20 +26,26 @@ class Alerts extends AlertsConfig
      *
      * @var array<string,string>
      */
-    public $classes = [
+    public array $classes = [
+        // Bootstrap classes
         'primary'   => 'primary',
-        'message'   => 'primary',
         'secondary' => 'secondary',
-        'notice'    => 'secondary',
         'success'   => 'success',
         'danger'    => 'danger',
+        'warning'   => 'warning',
+        'info'      => 'info',
+
+        // Additional log levels
+        'message'   => 'primary',
+        'notice'    => 'secondary',
         'error'     => 'danger',
-        'errors'    => 'danger',
         'critical'  => 'danger',
         'emergency' => 'danger',
-        'warning'   => 'warning',
         'alert'     => 'warning',
-        'info'      => 'info',
         'debug'     => 'info',
+
+        // Common framework keys
+        'messages' => 'primary',
+        'errors'   => 'danger',
     ];
 }
